Propose update to google-drive-folder-structure.md (L.Goenaga)
I propose a couple changes. To the Planning description I noted that it houses the staple Dig Ops documents: Brief, Plan, and Schedule.

I then propose merging the copywriting and creative folders into one execution folder. Basically, this execution folder houses all the paid, earned, owned, and social stuff we do for a client. It houses the stuff we route to a client for approval during the digital execution phase. I propose merging them because it seems to make sense to have the copy and creative in the same place, especially if we route them together like I tend to do with social ads.

I also propose adding a Reports folder, which will house the periodical reports we generate for the client, and also house that client's Dig Ops Tracker (which gets synced to Klipfolio via google drive).

Also updated it to say 5, instead of four, folders.